pyproject.toml
[tool.poetry]
name = "nonebot_plugin_remake"
version = "0.4.0"
description = "适用于 Nonebot2 的人生重开模拟器插件"
authors = ["meetwq <meetwq@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/noneplugin/nonebot-plugin-remake"
repository = "https://github.com/noneplugin/nonebot-plugin-remake"

[tool.poetry.dependencies]
python = "^3.8"
nonebot2 = "^2.2.0"
nonebot-plugin-alconna = ">=0.38.0,<1.0.0"
Pillow = "^10.0.0"

[tool.pyright]
pythonVersion = "3.8"
pythonPlatform = "All"
typeCheckingMode = "basic"

[tool.ruff]
line-length = 88
target-version = "py38"

[tool.ruff.lint]
select = ["E", "W", "F", "UP", "C", "T", "PYI", "PT", "Q"]
ignore = ["E402", "C901", "UP037"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
